+++ ApiTest/src/test/java/GetListTest.java
@@ -1,0 +1,12 @@
+import io.qameta.allure.Owner;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class GetListTest {
+    @DisplayName("Получение списка пользователей")
+    @Owner("Леонов Алексей Евгеньевич")
+    @Test
+    public void checkGoToVk() {
+
+    }
+}
