+++ UiTest/src/main/java/org/example/page/BasePage.java
@@ -1,0 +1,5 @@
+package org.example.page;
+
+public class BasePage extends BaseElement {
+
+}
