+++ UiTest/src/main/java/org/example/page/mail/LeftPanel.java
