+++ UiTest/src/main/java/org/example/page/mail/elements/MiniProfileSidebar.java
