+++ pom.xml
@@ -7,6 +7,11 @@
     <groupId>org.example</groupId>
     <artifactId>AutoTest</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+    <modules>
+        <module>UiTest</module>
+        <module>ApiTest</module>
+    </modules>
 
     <properties>
         <selenium-version>4.16.1</selenium-version>
@@ -22,11 +27,6 @@
     </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>org.seleniumhq.selenium</groupId>
-            <artifactId>selenium-java</artifactId>
-            <version>${selenium-version}</version>
-        </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
