--- src/main/java/org/example/page/BasePage.java
@@ -1,5 +1,0 @@
-package org.example.page;
-
-public class BasePage extends BaseElement {
-
-}
